refactor(#741): refactor locationController
